DavidK2's proofreading as part of CTC-349
Proofreading requested in https://kentico.atlassian.net/browse/CTC-349

Note: In the Learning Gatsby section, there are referenced sections of the Gatsby docs which don't appear to exist anymore. I don't know which particular parts are supposed to be referenced in the new structure so I'm leaving it as is for now: In particular, check out the “Guides”, API reference, and “Advanced Tutorials” sections in the sidebar.